CodeClash/backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary
python-dotenv
django-cors-headers
